+++ src/middleware.ts
@@ -1,7 +1,7 @@
+import { cookies } from 'next/headers';
 import { NextRequest, NextResponse } from 'next/server';
 
 import { ROOT, LOGIN, PUBLIC_ROUTES } from '@/lib/routes';
-import { cookies } from 'next/headers';
 
 const AUTH_COOKIE = 'access_token';
 
@@ -16,28 +16,14 @@ export async function middleware(request: NextRequest) {
       pathname.startsWith(route),
     );
 
-    console.log({
-      pathname,
-      isPublicRoute,
-      hasToken: !!token,
-    });
-
     // Redirect authenticated users away from login page
     if (token && pathname === LOGIN) {
-      console.log(
-        'Authenticated user trying to access login page, redirecting to root',
-      );
-
       const redirectUrl = new URL(ROOT, request.url);
       return NextResponse.redirect(redirectUrl);
     }
 
     // Redirect unauthenticated users to login if trying to access protected route
     if (!token && !isPublicRoute) {
-      console.log(
-        'Unauthenticated user trying to access protected route, redirecting to login',
-      );
-
       const loginUrl = new URL(LOGIN, request.url);
       loginUrl.searchParams.set('callbackUrl', pathname);
       return NextResponse.redirect(loginUrl);
